+++ docs/examples/example3-druid.js
@@ -45,8 +45,7 @@ ex.compute(context)
   .then(function(data) {
     // Log the data while converting it to a readable standard
     console.log(JSON.stringify(data.toJS(), null, 2));
-  })
-  .done();
+  });
 
 // ----------------------------------